moved list experiments to its own component
