+++ airavata-local-agent/renderer/lib/constants.js
@@ -1,0 +1,2 @@
+export const API_BASE_URL = 'https://md.cybershuttle.org/api';
+export const AUTH_BASE_URL = 'https://md.cybershuttle.org/auth';
